Improved tests and stability
I added some extra tests to catch the case of setting a container to hold
an empty container/assigning an empty container to a container. These revealed
the old massive memory allocation bug again, so I modified the
abstract_container%set method to avoid allocations which seem to trigger it.
I.e. allocating a temporary variable containing the type passed in to the set
method that time is another container.